Rearranged Underground2 support.

+++ GlobalLib/Support.Underground2/LoadData/Decompile/E_Materials.cs
@@ -15,7 +15,7 @@
             Utils.EA.Resolve.GetWindowTintString(CName);
             Core.Map.BinKeys[Utils.Bin.Hash(CName)] = CName;
 
-            var Class = new Class.Material(byteptr_t, CName, db);
+            var Class = new Class.Material((System.IntPtr)byteptr_t, CName, db);
             db.Materials.Classes[CName] = Class;
         }
     }
