+++ GlobalLib/Support.Underground2/LoadData/Decompile/E_PresetRides.cs
@@ -22,7 +22,7 @@
                 CName = Utils.EA.Resolve.GetPathFromCollection(CName);
                 Core.Map.BinKeys[Utils.Bin.Hash(CName)] = CName;
 
-                var Class = new Class.PresetRide(byteptr_t + offset, CName, db);
+                var Class = new Class.PresetRide((System.IntPtr)(byteptr_t + offset), CName, db);
                 db.PresetRides.Classes[CName] = Class;
             }
         }
