+++ GlobalLib/Support.Underground2/LoadData/Decompile/E_Tracks.cs
@@ -23,7 +23,7 @@
                 CName = Utils.EA.Resolve.GetPathFromCollection(CName);
                 Core.Map.BinKeys[Utils.Bin.Hash(CName)] = CName;
 
-                var Class = new Gameplay.Track(byteptr_t + offset, CName, db);
+                var Class = new Gameplay.Track((System.IntPtr)(byteptr_t + offset), CName, db);
                 db.Tracks.Classes[CName] = Class;
             }
         }
